sharing: refactor exception classes
