+++ classes/local/sharing/share_failed_exception.php
@@ -22,7 +22,7 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_collaborativefolders;
+namespace mod_collaborativefolders\local\sharing;
 
 defined('MOODLE_INTERNAL') || die();
